Add files via upload
This is what I was working on with Tatenda. The getquery() has been fixed but its missing Abuni's function. This version works better than phse3.1.py
Terms are working.
every other thing is not but some of the searches return 'no records found'.
What makes this better is there are no infinite loops stopping the functions from running.